+++ joulescope_ui/aspect_ratio_label.py
@@ -15,7 +15,7 @@
 """A Qt widget that displays an rescaled image to fit."""
 
 
-from PySide2 import QtWidgets, QtGui, QtCore
+from PySide6 import QtWidgets, QtGui, QtCore
 
 
 class AspectRatioLabel(QtWidgets.QLabel):