+++ joulescope_ui/widgets/developer/status_widget.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-from PySide2 import QtWidgets, QtCore
+from PySide6 import QtWidgets, QtCore
 from joulescope_ui.preferences_ui import widget_factory
 from joulescope.units import unit_prefix, three_sig_figs
 from joulescope_ui.ui_util import comboBoxConfig