+++ joulescope_ui/widgets/waveform/marker_area.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from PySide2 import QtWidgets, QtGui, QtCore
+from PySide6 import QtWidgets, QtGui, QtCore
 import pyqtgraph as pg
 import weakref
 from .signal import Signal