+++ joulescope_ui/widgets/waveform/signal_viewbox.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import pyqtgraph as pg
-from PySide2 import QtCore
+from PySide6 import QtCore
 from .scrollbar import WHEEL_TICK
 import logging